content/canvas/test/test_2d.line.defaults.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.line.defaults</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(ctx.lineWidth === 1, "ctx.lineWidth === 1");
ok(ctx.lineCap === 'butt', "ctx.lineCap === 'butt'");
ok(ctx.lineJoin === 'miter', "ctx.lineJoin === 'miter'");
ok(ctx.miterLimit === 10, "ctx.miterLimit === 10");

if (!_deferred) SimpleTest.finish();
});
</script>
